refactor: move configuration class to root directory
