+++ src/core/pages/page-routes.ts
@@ -4,7 +4,7 @@ import { Router, Request, Response, NextFunction } from "express";
 import * as express from "express";
 import { injectable } from "inversify";
 import { inject } from "inversify";
-import { ConfigurationService } from "../config/configuration";
+import { ConfigurationService } from "../../configuration";
 import HealthService, { HealthState } from "../health-service";
 
 @injectable()
