+++ src/sonar/sonar-status-emitter.ts
@@ -2,9 +2,9 @@ import { injectable, inject } from "inversify";
 import EventBus from "../core/event/event-bus";
 import { Events, GithubCheckRunWriteEvent } from "../core/event/event-model";
 import { ChecksCreateParams, ChecksCreateParamsOutputAnnotations } from "@octokit/rest";
-import { ConfigurationService } from "../core/config/configuration";
+import { ConfigurationService } from "../configuration";
 import { SonarWebhookEvent, QualityGateStatus } from "./model/sonar-wehook-event";
-import { SonarClient } from "./client/sonar-client";
+import SonarClient from "./client/sonar-client";
 import { LOGGER } from "../core/logger";
 import { Templates } from "../core/template/template-engine";
 import { TemplateEngine } from "../core/template/template-engine";
